agents/maverick-terminal-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maverick-terminal-agent"
version = "0.1.0"
description = "Maverick Terminal Agent - Terminal provisioning automation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Your Team", email = "your-email@example.com"}]
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0.0,<2",
    "typer>=0.9,<1",
    "pydantic>=2.0,<3",
    "pillow>=10,<11",
    "pytesseract>=0.3,<1",
]

[project.optional-dependencies]
ocr = [
    "easyocr>=1.7,<2",
    "pillow>=10,<11",
    "pytesseract>=0.3,<1",
]
dev = [
    "pytest>=7,<8",
    "pytest-asyncio>=0.21,<1",
    "ruff>=0.1,<1",
]

[project.scripts]
maverick = "maverick_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/maverick_agent"]
